Implement Admin Order Management and refractor some code (#55)
Introduces AdminOrder and CustomerOrder domain models, replacing the
previous generic Order and OrderItem entities. Adds admin order
functionality with new controllers, services, DTOs, and repository
interfaces/implementations. Updates user and customer order flows to use
the new models and decimal types for prices. Refactors infrastructure
and DI registration to support the new structure, and updates report
queries to use the new tables.

+++ src/Backend/OrderProcessing.Application/DTOs/Order/OrderItemDto.cs
@@ -2,5 +2,5 @@ namespace OrderProcessing.Application.DTOs.Order;
 
 public record OrderItemDto(
     int Quantity,
-    float UnitPrice
+    decimal UnitPrice
 );
